requirements.txt
numpy>=1.21.0
torch>=1.9.0
python-chess>=1.9.0
tqdm>=4.62.0
matplotlib>=3.4.0
pandas>=1.3.0
flask>=2.0.0 